pyproject.toml
[project]
name = "feishu-agent"
version = "0.1.0"
description = "Feishu Agent using MCP"
requires-python = ">=3.11"
dependencies = [
    "lark-oapi>=1.5.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "pydantic-settings>=2.0.0",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "pytest-asyncio>=0.23.0",
    "respx>=0.20.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]

